Fix cross compilation
In the proc_macro cfg(target_pointer_width) has the value of the host, not the target.
So instead, generate the string data for both the 32 and 64 bit, but have the generated
code select based on target_pointer_width.
This is not ideal from a compilation time point of view. Better suggestions welcome.

https://github.com/woboq/qmetaobject-rs/issues/19