feat(qttypes): Do more wrappers around QVariant
Wrapped converters to and constructors from all available types. The rest
of them are marked as TODO comments for future implementers.
Methods and constructors, which I'm not sure about whether they are needed
to be exposed in Rust at all, are marked with `TODO?:` prefix.

Constructors are reordered according to the official Qt documentation
for ease of code navigation. Return types are shortened to `Self`.

Fixed some misleading links to upstream docs in From<{i,u}{32,64}> impls.

Some wrappers are impossible to write at this point without introducing
QMetaType and its Type enum in this crate or its dependencies, as currently
it would create a circular deps between qmetaobject and qttypes crates.
Such hypothetical implementations are left commented out for future ideas.