README: Rewrite the headline (#165)
The %crate% is a crate which is a crate that is used...

Meh! No one reads such long headlines. And that is not even technically
correct, since qmetaobjec-rs is actually a bunch of crates under cargo
workspace umbrella.

New headline shamelessly mimics Rust because they have spent really good
amount of effort to make it short and attractive.

However, personally, I don't like the Rust headline that much. It says:

> A language empowering everyone
> **to build** reliable and efficient software.

Creating an app is not just about building it. But on the other hand,
Cargo is all about building Rust code. Further deployment process and
execution environment is up to a developer, as far as Rust is
concerned. I guess it's fine for them; but Qt's focus is a bit broader.