example(graph): Move main.qml into qrc
It probably was commented out because `qrc!()` didn't support pub fn
qualifier at the time of writing. Now it that it can be invoked from
main, we don't have to call it on each resource-dependent operation.
Also by baking main.qml into binary, we can get rid of runtime
dependency on CARGO_MANIFEST_DIR environment variable.

Possible improvement areas: qrc! doesn't support multiple identical
prefix declarations, so it is currently impossible write this:

```rust
qrc! {
    pub init_resources,
    "scenegraph/graph" {
        "shaders/noisy.vsh",
        "shaders/noisy.fsh",
        "shaders/line.vsh",
        "shaders/line.fsh",
    },
    "src" as "scenegraph/graph" {
        "main.qml"
    }
}

fn main() {
    view.set_source("qrc:/scenegraph/graph/main.qml".into());
}
```

There are no underlying technical limitations for it not to work though.
So, it might be implemented as an additional feature in the future.