Add tutorial on adding Rust wrappers for Qt C++ API
Adds set_flag as a separate method on Graph struct, to extend the
tutorial with fully usable wrapper and wrapper-specific documentation.

Fixes potential compilation issue of Graph example: cpp! with #include
was missing from main.rs, and the C++ code using QQuickItem
accidentally worked, because of #include declared in cpp! inside
another Rust file/module. But cpp crate does not specify in which order
content of cpp! content between files, so it should not be relied on.

Reorders #include lines in Graph example, such that external (Qt) stuff
is on top, and local imports follow below.

Resolves #169