build.rs: Use semver crate in qttypes
It is used in main qmetaobject crate anyway, so it's just more
convenient way to work with version object at no additional cost.

This line should behave the same as before, because Qt complies with
semver, so hopefully their versions are always going to be in
major.minor.patch format -- which is exactly the format used by Display
implementation of semver::Version.

    println!("cargo:VERSION={}", &qt_version);