Disable WebEngine example on Qt >= 6.0 && < 6.2
It was planned for Qt 6.2 LTS, according to Qt Roadmap for 2021.[1]

That's an awfully lots of changes simply to disable a single Qt module,
and it doesn't seem like disabling another one is going to take less
than this. I had to introduce build.rs to examples/webengine where it
is otherwise not needed.

On the bright side, cargo build and test commands once again work from
the root of the workspace, this time both with Qt 5 and 6.

[1]: https://www.qt.io/blog/qt-roadmap-for-2021